refactor: Change Plant ID type from int to long
This commit updates the data type of the `id` field in the `Plant` entity and associated methods from `int` to `long`.

- In `Plant.java`:
    - The `id` field is changed from `int` to `long`.
    - The getter `getId()` now returns `long`.
    - The setter `setId()` now accepts a `long` parameter.
- In `PlantRepository.java`:
    - In the `insert` method, the `id` returned by `plantDao.insert(plant)` (which is a `long`) is now directly assigned to `plant.setId()` without casting to `int`.

+++ app/src/main/java/de/oabidi/pflanzenbestandundlichttest/PlantRepository.java
@@ -56,7 +56,7 @@ public class PlantRepository {
     public Future<?> insert(Plant plant, Runnable callback) {
         return PlantDatabase.databaseWriteExecutor.submit(() -> {
             long id = plantDao.insert(plant);
-            plant.setId((int) id);
+            plant.setId(id);
             if (callback != null) {
                 mainHandler.post(callback);
             }
